+++ facdb/__init__.py
@@ -2,7 +2,7 @@ from .geocode.function1A import Function1A
 from .geocode.function1B import Function1B
 from .geocode.functionBL import FunctionBL
 from .geocode.functionBN import FunctionBN
-from .geocode.parseAddress import ParseAddress
+from .geocode.parseAddress import ParseAddress, UseAirportName
 from .utility.executeSQL import ExecuteSQL
 from .utility.export import Export
 from .utility.prepare import Prepare
